+++ lib/din.dart
@@ -11,6 +11,7 @@ export 'src/clients/gateway_client.dart'
         GatewayIdentify,
         GatewayIdentifyStrategy,
         PresenceStatus;
+export 'src/clients/gateway_events.dart' show GatewayEvents;
 export 'src/clients/http_client.dart' show HttpClient, HttpClientException;
 export 'src/clients/rest_client.dart' show RestClient;
 export 'src/clients/ws_client.dart' show WebSocketFactory, WebSocketClient;
@@ -18,7 +19,8 @@ export 'src/schema/resources/channel.dart' show ChannelsResource;
 export 'src/schema/resources/user.dart' show UsersResource;
 export 'src/schema/structures/channel.dart' show Channel, ChannelType;
 export 'src/schema/structures/gateway.dart'
-    show GatewayDispatch, Gateway, GatewayOpcode, GatewayReady;
+    show GatewayDispatch, Gateway, GatewayOpcode;
 export 'src/schema/structures/message.dart' show Message;
+export 'src/schema/structures/ready.dart' show Ready;
 export 'src/schema/structures/user.dart' show User;
 export 'src/user_agent.dart' show UserAgent;