+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0-alpha+7
+
+* Split most of the Gateway events into `GatewayClient.events`.
+* Renamed `GatewayReady` to `Ready`, and moved to `GatewayEvents.ready`.
+
 ## 0.1.0-alpha+6
 
 * Added `FakeHttpClient` in `platform/test.dart` for simple testing.